Moved Header and Footer into components
