+++ context/app/static/js/components/App.jsx
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 import Container from '@material-ui/core/Container';
 import Providers from './Providers';
 import Routes from './Routes';
-import Footer from '../Footer';
-import Header from '../Header';
+import Footer from './Footer';
+import Header from './Header';
 
 function App(props) {
   const { flaskData } = props;
