+++ context/app/static/js/index.jsx
@@ -7,8 +7,8 @@ import 'typeface-inter';
 import theme from './theme';
 import './index.css';
 import App from './components/App';
-import Footer from './Footer';
-import Header from './Header';
+import Footer from './components/Footer';
+import Header from './components/Header';
 
 const urlPath = window.location.pathname;
